Move workflow rewrite out of v2 package.
